Golden directory selection from the viewer.  still need a back button
