+++ lib/utils/serve-viewer.js
@@ -8,7 +8,7 @@ const chalk = require("chalk");
 const {resolveNodeModuleFile} = require('./index');
 const {readTestFile} = require('../utils/format-tests');
 const {writeGoldensManifest} = require('../utils/generate-screenshot');
-const {getGoldenImagesForViewer} = require('../utils/get-goldens-for-viewer');
+const {getGoldenImagesForViewer,getGoldenDirectoriesForViewer} = require('../utils/get-goldens-for-viewer');
 
 exports.serveViewer = (conf, resultsPath) => {
   const projectDir = conf.projectDir;
@@ -31,17 +31,28 @@ exports.serveViewer = (conf, resultsPath) => {
     next();
   });
 
-  /* serve the results json file or golden images when requested */
-  app.get('/results.json', (req, res) => {
+  /* serve the viewer config */
+  app.get('/viewer-config', (req, res) => {
     res.header('Content-Type', 'application/json');
-    let result;
-    if (conf.golden) {
-      const imageArr = getGoldenImagesForViewer(conf);
-      result = {golden:true, images:imageArr};
-    } else {
-      result = resultsJSON;
-    }
-    res.send(JSON.stringify(result));
+    res.send(JSON.stringify({golden:conf.golden}));
+  });
+
+  /* serve the results json file or golden images when requested */
+  app.get('/test-results', (req, res) => {
+    res.header('Content-Type', 'application/json');
+    res.send(JSON.stringify(resultsJSON));
+  });
+
+  /* serve the results json file or golden images when requested */
+  app.get('/test-results/*', (req, res) => {
+    res.header('Content-Type', 'application/json');
+    res.send(JSON.stringify({images:getGoldenImagesForViewer(conf, req.params[0])}));
+  });
+
+  /* serve the list of golden directories */
+  app.get('/golden-dirs', (req, res) => {
+    res.header('Content-Type', 'application/json');
+    res.send(JSON.stringify(getGoldenDirectoriesForViewer(conf)));
   });
 
   /* serve the test images when requested */
